requirements.txt
django-allauth==0.37.1